extract getGitHubItems
